requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
gspread>=5.10.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
st-gsheets-connection>=0.0.4
openpyxl>=3.1.0
xlrd>=2.0.0
